Fix build error on MAcosX
git-svn-id: https://llvm.org/svn/llvm-project/vmkit/branches/release_027@100108 91177308-0d34-0410-b5e6-96231b3b80d8

+++ lib/Mvm/MMTk/MvmGC.cpp
@@ -31,7 +31,7 @@ extern "C" void* gcmalloc(size_t sz, void* _VT) {
   return res;
 }
 
-extern "C" void* gcmallocUnresolved(size_t sz, VirtualTable* VT) {
+extern "C" void* gcmallocUnresolved(uint32_t sz, VirtualTable* VT) {
   gc* res = (gc*)gcmalloc(sz, VT);
   if (VT->destructor)
     mvm::Thread::get()->MyVM->addFinalizationCandidate(res);
